python/requirements.txt
cloudzeta-sdk==0.6.73